armv8: Handle EL2 Host mode
On implementations that support VHE, the layout of the CPTR_EL2
register depends on whether HCR_EL2.E2H is set.  Check this bit
and and set the aprropriate bits to enable access to the FP/SIMD
registers.  This allows U-Boot to run on systems that pass
control to U-Boot in EL2 with EL2 Host mode enabled such as
machine with Apple's M1 SoC.

Signed-off-by: Mark Kettenis <kettenis@openbsd.org>